+++ packages/base-editor/src/components/InjectView.jsx
@@ -1,6 +1,13 @@
+import { View } from "@elf-framework/ui";
+
 import { useEditor } from "./editor/Editor";
 
-export function InjectView({ views = [], groups = [] }) {
+export function InjectView({
+  views = [],
+  groups = [],
+  as = "div",
+  style = {},
+}) {
   const editor = useEditor();
 
   // views, groups 에 있던 컴포넌트를 하나의 배열로 변환
@@ -15,5 +22,11 @@ export function InjectView({ views = [], groups = [] }) {
     .flat(Infinity)
     .filter(Boolean);
 
-  return <>{list}</>;
+  // console.log("list", groups, views, list);
+
+  return (
+    <View as={as} style={style}>
+      {list}
+    </View>
+  );
 }