+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @elf-framework/design-system
 
+## 2.0.10
+
+### Patch Changes
+
+- modify base ui
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@2.0.10
+
 ## 1.1.1
 
 ### Patch Changes